requirements.txt
requests>=2.31.0
loguru>=0.7.2
sqlalchemy>=2.0.0
schedule>=1.2.0
pandas>=2.0.0
newsapi-python>=0.2.7
feedparser>=6.0.0
beautifulsoup4>=4.12.0
rapidfuzz>=3.6.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# Testing
pytest>=9.0.0
httpx>=0.27.0
